requirements.txt
google-cloud-aiplatform>=1.26.0
psycopg-binary>=2.9
python-dotenv>=0.21.0